React package nests metadata under spaceInfo, other flows don’t. So support both. Also converted to .ts.
